♻️ Restructure explorer architecture boundaries
