+++ crates/core/src/explorer/mod.rs
@@ -1,1 +1,2 @@
 pub mod jobs;
+pub mod projection;
