+++ crates/core/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod explorer;
 mod ipc;
 
-pub use explorer::jobs;
+pub use explorer::{jobs, projection};
 pub use ipc::directory;
