+++ package.json
@@ -4,12 +4,12 @@
   "description": "Windows-first file explorer built with a Rust core and a thin Tauri renderer.",
   "type": "module",
   "scripts": {
-    "dev": "vite dev",
-    "build": "svelte-kit sync && vite build",
-    "preview": "vite preview",
-    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "tauri": "tauri"
+    "dev": "bun node_modules/vite/bin/vite.js dev",
+    "build": "bun node_modules/@sveltejs/kit/svelte-kit.js sync && bun node_modules/vite/bin/vite.js build",
+    "preview": "bun node_modules/vite/bin/vite.js preview",
+    "check": "bun node_modules/@sveltejs/kit/svelte-kit.js sync && bun node_modules/svelte-check/bin/svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "bun node_modules/@sveltejs/kit/svelte-kit.js sync && bun node_modules/svelte-check/bin/svelte-check --tsconfig ./tsconfig.json --watch",
+    "tauri": "bun node_modules/@tauri-apps/cli/tauri.js"
   },
   "license": "MIT",
   "dependencies": {
